src/main/java/ua/nure/ipz/zoo/entity/Animal.java
package ua.nure.ipz.zoo.entity;

import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.Lob;
import javax.persistence.ManyToOne;

import ua.nure.ipz.zoo.util.DomainEntity;

@Entity
public class Animal extends DomainEntity {
    @ManyToOne
    private Aviary aviary;
    private String name;
    private String type;
    private String imageUrl;

    @Column(columnDefinition = "clob")
    @Lob
    private String description;

    public Animal() {
    }

    public Animal(String name, String type, String imageUrl) {
        this.name = name;
        this.type = type;
        this.imageUrl = imageUrl;
    }

    public Aviary getAviary() {
        return aviary;
    }

    public void setAviary(Aviary aviary) {
        this.aviary = aviary;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getType() {
        return type;
    }

    public void setType(String type) {
        this.type = type;
    }

    public String getImageUrl() {
        return imageUrl;
    }

    public void setImageUrl(String imageUrl) {
        this.imageUrl = imageUrl;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }
}
